% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abb2fips.R
\name{abb2fips}
\alias{abb2fips}
\title{abb2fips()}
\usage{
abb2fips(stateabb)
}
\arguments{
\item{stateabb}{state abb as two character abb i.e. "CA", "TX", or "NY"}
}
\description{
Get fips code from state abbreviation
}
