% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abb2state.R
\name{abb2state}
\alias{abb2state}
\title{abb2state()}
\usage{
abb2state(stateabb)
}
\arguments{
\item{stateabb}{state abb as two character abb i.e. "CA", "TX", or "NY"}
}
\description{
Get state from state abbreviation
}
