% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city2zips.R
\name{city2zips}
\alias{city2zips}
\title{city2zips()}
\usage{
city2zips(city, state = NA)
}
\arguments{
\item{city}{city as city name i.e. "Austin" or "Los Angeles" any capitalization but including spaces}

\item{state}{state to return zips for as two character abb i.e. "CA", "TX", or "NY"}
}
\description{
Filter zipcodes to certain state
}
