% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county2zips.R
\name{county2zips}
\alias{county2zips}
\title{county2zips()}
\usage{
county2zips(
  county,
  state,
  cfips = NA,
  append_county = TRUE,
  append_state = TRUE
)
}
\arguments{
\item{county}{city as city name i.e. "Austin" or "Los Angeles" any capitalization but including spaces}

\item{state}{state to return zips for as two character abb i.e. "CA", "TX", or "NY"}

\item{cfips}{optional county as 5 digit fips code i.e. "08005" for Arapahoe County, Colorado}

\item{append_county}{append county name to dataset as column countyname}

\item{append_state}{append state name to dataset as column statename}
}
\description{
Filter zipcodes to certain state
}
