% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcty.R
\name{fcty}
\alias{fcty}
\title{fcty()}
\usage{
fcty(state, cname, tfwformat = TRUE)
}
\arguments{
\item{state}{state to return counties for, may be two character "CO" or full name "Colorado"}

\item{cname}{countyname to find}

\item{tfwformat}{return simple format}
}
\description{
Get all counties for a state
}
