% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state2counties.R
\name{state2counties}
\alias{state2counties}
\title{state2counties()}
\usage{
state2counties(state)
}
\arguments{
\item{state}{state to return counties for, may be two character "CO" or full name "Colorado"}
}
\description{
Get all counties for a state
}
