% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state2fips.R
\name{state2fips}
\alias{state2fips}
\title{state2fips()}
\usage{
state2fips(statename)
}
\arguments{
\item{stateabb}{state name in full i.e. "California", "New York"}
}
\description{
Get fips from state name
}
