% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state2zips.R
\name{state2zips}
\alias{state2zips}
\title{state2zips()}
\usage{
state2zips(state)
}
\arguments{
\item{state}{state to return zips for as two character abb i.e. "CA", "TX", or "NY"}
}
\description{
Filter zipcodes to certain state
}
