% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_filt.R
\name{zip_filt}
\alias{zip_filt}
\title{zip_filt()}
\usage{
zip_filt(indata, state, zipcol = "zipcode")
}
\arguments{
\item{indata}{data.frame object to filter}

\item{state}{state as two character abb i.e. "CA", "TX", or "NY"}

\item{zipcol}{column name of zipcode column in indata, default="zipcode"}
}
\description{
Filter zipcodes to certain state
}
